Drop prettier from workspace
