+++ applications/browser-extension/package.json
@@ -278,7 +278,6 @@
     "node-polyfill-webpack-plugin": "^4.0.0",
     "otpauth": "^9.3.4",
     "path-browserify": "^1.0.1",
-    "prettier": "3.1.0",
     "react-beautiful-dnd-next": "^11.0.5",
     "react-refresh-typescript": "^2.0.9",
     "react-select-event": "^5.3.0",
